update project for zuul v3
This patch updates the project for zuul v3 by:
- Adding the lower-constraints.txt which defines the lower limits for
dependencies and a corresponding tox target and gate job.
- Updating tox.ini to use global constraints rather than installing
neutron from git as well as switching over to stestr for running UTs.
- Updating the requirements to pull in the latest neutron-lib as well
as define neutron as required by gate jobs.
- Updating test-requirementst to include stestr.
- Updating .zuul.yaml to use the neutron templates for python, lower
constraints and requirements jobs. This ensures neutron is pulled from
source in the gate.

Change-Id: I3dd8138d3b46f76a75e354b4eb3b96cca782ed8e